+++ example/android/app/src/main/kotlin/com/google/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.google.maps
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}